+++ cpp/include/cudf/aggregation.hpp
@@ -43,6 +43,12 @@ enum class udf_type : bool {
    PTX
 };
 
+// @brief Enum to describe include nulls or exclude nulls in an aggregation
+enum class include_nulls : bool {
+   YES, 
+   NO
+};
+
 /// Factory to create a SUM aggregation
 std::unique_ptr<aggregation> make_sum_aggregation();
 
@@ -117,7 +123,8 @@ std::unique_ptr<aggregation> make_argmin_aggregation();
  * 
  * `nunique` returns the number of unique elements.
 */
-std::unique_ptr<aggregation> make_nunique_aggregation();
+std::unique_ptr<aggregation>
+make_nunique_aggregation(include_nulls _include_nulls = include_nulls::NO);
 /**
  * @brief Factory to create a aggregation base on UDF for PTX or CUDA
  *